Complete functional integration for Player and UI
Refactor some functionality into GameController class.
Player and UI classes now reference the GameController for money and
timer values.

Player input is not yet being paused when the game is paused. This is waiting on the input controller to be implemented.

+++ Assets/Scripts/LevelData.cs
@@ -11,7 +11,7 @@ public class LevelData
     public int MoneyGoal;
 
     // The level's timer in seconds
-    public int RemainingTime;
+    public float RemainingTime;
 
     // Hold the map's tile layout
     public Texture2D TileLayout;
